@@ -15,7 +15,7 @@ type geoipFilter struct{}
 
 func (gf *geoipFilter) Match(ip net.IP) bool {
 	record, _ := mmdb.Instance().Country(ip)
-	return record.Country.IsoCode != "CN" && record.Country.IsoCode != ""
+	return record.Country.IsoCode != "CN" && !ip.IsPrivate()
 }
 
 type ipnetFilter struct {
