@@ -10,7 +10,24 @@ import (
 
 type Listener struct {
 	listener net.Listener
+	addr     string
 	closed   bool
+}
+
+// RawAddress implements C.Listener
+func (l *Listener) RawAddress() string {
+	return l.addr
+}
+
+// Address implements C.Listener
+func (l *Listener) Address() string {
+	return l.listener.Addr().String()
+}
+
+// Close implements C.Listener
+func (l *Listener) Close() error {
+	l.closed = true
+	return l.listener.Close()
 }
 
 func New(addr string, in chan<- C.ConnContext) (*Listener, error) {
@@ -30,6 +47,7 @@ func NewWithAuthenticate(addr string, in chan<- C.ConnContext, authenticate bool
 
 	hl := &Listener{
 		listener: l,
+		addr:     addr,
 	}
 	go func() {
 		for {
@@ -45,13 +63,4 @@ func NewWithAuthenticate(addr string, in chan<- C.ConnContext, authenticate bool
 	}()
 
 	return hl, nil
-}
-
-func (l *Listener) Close() {
-	l.closed = true
-	l.listener.Close()
-}
-
-func (l *Listener) Address() string {
-	return l.listener.Addr().String()
 }
