@@ -9,7 +9,24 @@ import (
 
 type Listener struct {
 	listener net.Listener
+	addr     string
 	closed   bool
+}
+
+// RawAddress implements C.Listener
+func (l *Listener) RawAddress() string {
+	return l.addr
+}
+
+// Address implements C.Listener
+func (l *Listener) Address() string {
+	return l.listener.Addr().String()
+}
+
+// Close implements C.Listener
+func (l *Listener) Close() error {
+	l.closed = true
+	return l.listener.Close()
 }
 
 func New(addr string, in chan<- C.ConnContext) (*Listener, error) {
@@ -19,6 +36,7 @@ func New(addr string, in chan<- C.ConnContext) (*Listener, error) {
 	}
 	rl := &Listener{
 		listener: l,
+		addr:     addr,
 	}
 
 	go func() {
@@ -35,15 +53,6 @@ func New(addr string, in chan<- C.ConnContext) (*Listener, error) {
 	}()
 
 	return rl, nil
-}
-
-func (l *Listener) Close() {
-	l.closed = true
-	l.listener.Close()
-}
-
-func (l *Listener) Address() string {
-	return l.listener.Addr().String()
 }
 
 func handleRedir(conn net.Conn, in chan<- C.ConnContext) {
