@@ -10,7 +10,30 @@ import (
 
 type Listener struct {
 	listener net.Listener
+	addr     string
 	closed   bool
+}
+
+// RawAddress implements C.Listener
+func (l *Listener) RawAddress() string {
+	return l.addr
+}
+
+// Address implements C.Listener
+func (l *Listener) Address() string {
+	return l.listener.Addr().String()
+}
+
+// Close implements C.Listener
+func (l *Listener) Close() error {
+	l.closed = true
+	return l.listener.Close()
+}
+
+func (l *Listener) handleTProxy(conn net.Conn, in chan<- C.ConnContext) {
+	target := socks5.ParseAddrToSocksAddr(conn.LocalAddr())
+	conn.(*net.TCPConn).SetKeepAlive(true)
+	in <- inbound.NewSocket(target, conn, C.TPROXY)
 }
 
 func New(addr string, in chan<- C.ConnContext) (*Listener, error) {
@@ -32,6 +55,7 @@ func New(addr string, in chan<- C.ConnContext) (*Listener, error) {
 
 	rl := &Listener{
 		listener: l,
+		addr:     addr,
 	}
 
 	go func() {
@@ -48,19 +72,4 @@ func New(addr string, in chan<- C.ConnContext) (*Listener, error) {
 	}()
 
 	return rl, nil
-}
-
-func (l *Listener) Close() {
-	l.closed = true
-	l.listener.Close()
-}
-
-func (l *Listener) Address() string {
-	return l.listener.Addr().String()
-}
-
-func (l *Listener) handleTProxy(conn net.Conn, in chan<- C.ConnContext) {
-	target := socks5.ParseAddrToSocksAddr(conn.LocalAddr())
-	conn.(*net.TCPConn).SetKeepAlive(true)
-	in <- inbound.NewSocket(target, conn, C.TPROXY)
 }
