@@ -11,7 +11,24 @@ import (
 
 type UDPListener struct {
 	packetConn net.PacketConn
+	addr       string
 	closed     bool
+}
+
+// RawAddress implements C.Listener
+func (l *UDPListener) RawAddress() string {
+	return l.addr
+}
+
+// Address implements C.Listener
+func (l *UDPListener) Address() string {
+	return l.packetConn.LocalAddr().String()
+}
+
+// Close implements C.Listener
+func (l *UDPListener) Close() error {
+	l.closed = true
+	return l.packetConn.Close()
 }
 
 func NewUDP(addr string, in chan<- *inbound.PacketAdapter) (*UDPListener, error) {
@@ -22,6 +39,7 @@ func NewUDP(addr string, in chan<- *inbound.PacketAdapter) (*UDPListener, error)
 
 	rl := &UDPListener{
 		packetConn: l,
+		addr:       addr,
 	}
 
 	c := l.(*net.UDPConn)
@@ -58,15 +76,6 @@ func NewUDP(addr string, in chan<- *inbound.PacketAdapter) (*UDPListener, error)
 	}()
 
 	return rl, nil
-}
-
-func (l *UDPListener) Close() error {
-	l.closed = true
-	return l.packetConn.Close()
-}
-
-func (l *UDPListener) Address() string {
-	return l.packetConn.LocalAddr().String()
 }
 
 func handlePacketConn(pc net.PacketConn, in chan<- *inbound.PacketAdapter, buf []byte, lAddr *net.UDPAddr, rAddr *net.UDPAddr) {
